Replace usage of f64 by f32

+++ src/node_trait.rs
@@ -2,5 +2,5 @@ use crate::node_values::NodeValues;
 
 pub(crate) trait NodeTrait {
     fn get_id(&self) -> usize;
-    fn process(&mut self, node_values: &NodeValues) -> f64;
+    fn process(&mut self, node_values: &NodeValues) -> f32;
 }
